backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aicg-platform-backend"
version = "1.0.0"
description = "AICG内容分发平台 - 后端服务"
authors = [
    { name = "AICG Platform Team", email = "team@aicg-platform.com" },
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.12.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "minio>=7.2.0",
    "structlog>=23.2.0",
    "loguru>=0.7.2",
    "httpx>=0.25.0",
    "python-magic>=0.4.27",
    "ffmpeg-python>=0.2.0",
    "pillow>=10.0.0",
    "openai>=1.3.0",
    "cryptography>=41.0.0",
    "prometheus-client>=0.18.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.42b0",
    "opentelemetry-instrumentation-redis>=0.42b0",
    "psutil>=5.9.0",
    "pytz>=2023.3",
    "python-dateutil>=2.8.2",
    "email-validator>=2.0.0",
    "jinja2>=3.1.2",
    "colorama>=0.4.6",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.4.0",
    "aiosqlite>=0.19.0",
    "factory-boy>=3.3.0",
    "faker>=19.12.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "aiosqlite>=0.19.0",
    "factory-boy>=3.3.0",
    "faker>=19.12.0",
]

[project.urls]
Homepage = "https://github.com/your-org/aicg-platform"
Documentation = "https://docs.aicg-platform.com"
Repository = "https://github.com/your-org/aicg-platform.git"
"Bug Tracker" = "https://github.com/your-org/aicg-platform/issues"

[project.scripts]
aicg-server = "src.main:main"
aicg-worker = "src.workers.base:start_worker"

[tool.hatch.version]
path = "src/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/migrations",
    "/scripts",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

# Black配置
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort配置
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]
known_third_party = ["fastapi", "sqlalchemy", "celery", "redis", "minio"]

# mypy配置
[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "redis.*",
    "minio.*",
    "ffmpeg.*",
    "magic.*",
    "structlog.*",
    "loguru.*",
    "prometheus_client.*",
    "opentelemetry.*",
    "factory_boy.*",
    "faker.*",
]
ignore_missing_imports = true

# pytest配置
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "contract: marks tests as contract tests",
    "e2e: marks tests as end-to-end tests",
]

# Coverage配置
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/migrations/*",
    "*/scripts/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"